lab2/permissions/requirements.txt
flask
jsonschema
requests
mysql-connector-python
